display "view artist" in case of music in actions menu instead of "view channel"
