+++ netlify/edge-functions/opengraph.ts
@@ -12,7 +12,7 @@ export default async (request: Request, context: Context) => {
 
   const response = await context.next();
   const page = await response.text();
-  const instance = 'https://invidious.jing.rocks';
+  const instance = 'https://inv.qilk.de';
   const data = await fetch(instance + '/api/v1/videos/' + id).then(res => res.json());
   const music = data.author.endsWith(' - Topic') ? 'https://wsrv.nl?w=180&h=180&fit=cover&url=' : '';
   const thumbnail = music + data.videoThumbnails.find((v: { quality: string }) => v.quality === 'medium').url;
