+++ netlify/edge-functions/public.ts
@@ -1,31 +1,7 @@
 // handles upcoming query & public collection requests to restore stream lists from any state
 
 import { Config } from '@netlify/edge-functions';
-
-
-function convertSStoHHMMSS(seconds: number): string {
-  if (seconds < 0) return '';
-  const hh = Math.floor(seconds / 3600);
-  seconds %= 3600;
-  const mm = Math.floor(seconds / 60);
-  const ss = Math.floor(seconds % 60);
-  let mmStr = String(mm);
-  let ssStr = String(ss);
-  if (mm < 10) mmStr = '0' + mmStr;
-  if (ss < 10) ssStr = '0' + ssStr;
-  return (hh > 0 ?
-    hh + ':' : '') + `${mmStr}:${ssStr}`;
-}
-
-
-const instanceArray: string[] = [
-  'https://invidious.perennialte.ch/api/v1/videos/',
-  'https://invidious.jing.rocks/api/v1/videos/',
-  'https://inv.tux.pizza/api/v1/videos/',
-  'https://invidious.nerdvpn.de/api/v1/videos/',
-  'https://invidious.privacyredirect.com/api/v1/videos/',
-  'https://inv.nadeko.net/api/v1/videos/'
-];
+import { convertSStoHHMMSS, instanceArray } from './commons';
 
 const getIndex = () => Math.floor(Math.random() * instanceArray.length);
